Refactor level screens to fix error
