+++ core/src/asu/gunma/ui/screen/game/GameScreen.java
@@ -1,13 +1,10 @@
 package asu.gunma.ui.screen.game;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Preferences;
 import com.badlogic.gdx.Screen;
+import com.badlogic.gdx.audio.Music;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.Texture;
@@ -24,9 +21,13 @@ import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
-import com.badlogic.gdx.audio.Music;
 import com.badlogic.gdx.utils.Align;
-import asu.gunma.DatabaseInterface.*;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
+
+import asu.gunma.DatabaseInterface.DbInterface;
 import asu.gunma.DbContainers.VocabWord;
 import asu.gunma.speech.ActionResolver;
 import asu.gunma.ui.screen.menu.MainMenuScreen;
@@ -126,8 +127,9 @@ public class GameScreen implements Screen {
     Random rand = new Random();
 
     Preferences prefs;
+    int levelNumber;
 
-    public GameScreen(Game game, ActionResolver speechGDX, Music music, DbInterface dbCallback, Screen previous, ArrayList<VocabWord> activeList, Preferences prefs, GameAssets gameAssets) {
+    public GameScreen(Game game, ActionResolver speechGDX, Music music, DbInterface dbCallback, Screen previous, ArrayList<VocabWord> activeList, Preferences prefs, GameAssets gameAssets, int levelNumber) {
         this.game = game;
         this.prefs = prefs;
         this.speechGDX = speechGDX;
@@ -136,6 +138,7 @@ public class GameScreen implements Screen {
         this.gameMusic = music;
         this.activeVList = activeList;
         this.gameAssets = gameAssets;
+        this.levelNumber = levelNumber;
         gameMusic = Gdx.audio.newMusic(Gdx.files.internal(gameAssets.introMusicPath));
         gameMusic.setLooping(false);
         gameMusic.setVolume(masterVolume);
@@ -162,7 +165,7 @@ public class GameScreen implements Screen {
         this.livesDrawer = new LivesDrawer(this.batch);
 
         // Animation initializations
-        this.onionWalkAnimation = new Animator(gameAssets.onionWalkAnimationPath, 4, 2, 0.1f);
+        this.onionWalkAnimation = new Animator(gameAssets.frenemyWalkAnimationPathPerLevel[levelNumber - 1], 4, 2, 0.1f);
         this.gunmaWalkAnimation = new Animator(gameAssets.gunmaWalkAnimationPath, 8, 1, 0.1f);
 
         // Game feedback
@@ -218,6 +221,7 @@ public class GameScreen implements Screen {
         parameter2.color = Color.BLACK;
 //        font2 = generator.generateFont(parameter2);
         font2 = gameAssets.getFont();
+        font2.setColor(0, 0, 0, 1);
 
         //Alignment and Text Wrapping for Vocab Word
         displayWordLayout = new GlyphLayout();
@@ -285,7 +289,6 @@ public class GameScreen implements Screen {
                 gameMusic.setVolume(masterVolume);
                 gameMusic.play();
                 game.setScreen(new MainMenuScreen(game, speechGDX,  gameMusic, dbCallback,activeVList, prefs, gameAssets));
-                previousScreen.dispose();
                 dispose(); // dispose of current GameScreen
             }
         });
@@ -381,15 +384,16 @@ public class GameScreen implements Screen {
             this.walkOntoScreenFromRight(delta);
         } else {
             speechGDX.stopRecognition();
-
-            if(win) {
-                font2.draw(batch, "You Win!", 450, 380);
-                // batch.draw(supergunma, 70, 10 + this.SCREEN_BOTTOM_ADJUST);
+            int numStars = 0;
+            if(score >= gameAssets.threeStarRequirement[levelNumber - 1]) {
+                numStars = 3;
+            } else if(score >= gameAssets.twoStarRequirement[levelNumber - 1]) {
+                numStars = 2;
+            } else if(score >= gameAssets.oneStarRequirement[levelNumber - 1]) {
+                numStars = 1;
             }
-            else{
-                font2.draw(batch, "You Lose!", 450, 380);
-                batch.draw(this.gunmaFaintedSprite, 70, 10 + this.SCREEN_BOTTOM_ADJUST);
-            }
+            gameAssets.setLevelStars(levelNumber, numStars);
+            addScore(numStars);
         }
 
         if(correctDisplayTimer > 0) { this.correctAnswerGraphic();}
@@ -445,7 +449,8 @@ public class GameScreen implements Screen {
             tmp.flip(true, false);
             batch.draw(tmp, this.enemyPosition, 40 + this.SCREEN_BOTTOM_ADJUST);
             tmp.flip(true, false);
-            this.enemyPosition -= 1.15;
+            // enemy speed
+            this.enemyPosition -= gameAssets.frenemySpeed[levelNumber - 1];
             if (this.enemyPosition < 100) {
                 this.takeDamage();
             }
@@ -496,5 +501,51 @@ public class GameScreen implements Screen {
         Random rand = new Random();
         int randomInt = rand.nextInt(size - 1);
         return randomInt;
+    }
+
+    private void addScore(int numStars) {
+        Table table = new Table();
+        table.setPosition(Gdx.graphics.getWidth()/2, Gdx.graphics.getHeight()/2);
+        Texture stars = new Texture(gameAssets.getStarPath(numStars));
+
+        batch.draw(this.gunmaFaintedSprite, 70, 10 + this.SCREEN_BOTTOM_ADJUST);
+        batch.draw(stars, Gdx.graphics.getWidth()/2 - stars.getWidth()/4, Gdx.graphics.getHeight()/2 + stars.getHeight()/2, stars.getWidth()/2, stars.getHeight()/2);
+        TextButton.TextButtonStyle textButtonStyle = new TextButton.TextButtonStyle();
+        textButtonStyle.pressedOffsetX = 1;
+        textButtonStyle.pressedOffsetY = -1;
+        textButtonStyle.font = font2;
+        textButtonStyle.fontColor = Color.BLACK;
+
+        Label.LabelStyle headingStyle = new Label.LabelStyle(font2, Color.BLACK);
+
+        Label heading = new Label(gameAssets.getResourceBundle().getString("YourScore") + " " + score, headingStyle);
+        Skin skin = gameAssets.getColorSkin(gameAssets.color2, "color2");
+        textButtonStyle.up = skin.newDrawable("color2", gameAssets.color2);
+
+        TextButton continueButton = new TextButton(gameAssets.getResourceBundle().getString("Continue"), textButtonStyle);
+
+        heading.setFontScale(2);
+
+
+        continueButton.addListener(new ClickListener() {
+            public void clicked(InputEvent event, float x, float y) {
+                speechGDX.stopRecognition();
+                isPaused = true;
+                gameMusic.dispose();
+                gameMusic = Gdx.audio.newMusic(Gdx.files.internal(gameAssets.introMusicPath));
+                gameMusic.setLooping(false);
+                gameMusic.setVolume(masterVolume);
+                gameMusic.play();
+                game.setScreen(new MainMenuScreen(game, speechGDX,  gameMusic, dbCallback, activeVList, prefs, gameAssets));
+                dispose(); // dispose of current GameScreen
+            }
+        });
+
+        continueButton.pad(15);
+        table.add(heading).padBottom(20);
+        table.row();
+        table.add(continueButton);
+
+        stage.addActor(table);
     }
 }
